requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
openai==0.27.0
python-dotenv==1.0.0
WeRoBot==1.13.1
